requirements.txt
# -*- coding: utf-8 -*-
# 工程预装模块
# 需要额外的python包，可直接在文件后面添加
# 请确保指定的包和版本号，可通过pip安装

# blueapps requirement
Django==2.2.6
mysqlclient==1.4.4
django-mysql==3.9.0
MarkupSafe==1.1.1
# Mako==1.0.6
requests==2.22.0
celery==3.1.25
django-celery==3.2.1
python-json-logger==0.1.7
whitenoise==3.3.0
six==1.11.0
djangorestframework==3.12.4
django-cors-middleware==1.5.0
django-cors-headers==3.2.1

django-filter==2.4.0
djangorestframework-jwt==1.11.0
xlsxwriter==1.4.3
IPy==1.1
psutil==5.8.0
ipaddr==2.2.0
xlrd==1.2.0
redis==2.10.6
